move mobile nav to own component
