+++ frontend/src/assets/stylesheets/base/1_globals.scss
@@ -7,7 +7,7 @@ body {
 }
 
 main {
-  margin: 0 auto;
+  margin: 2em auto;
 }
 
 @media screen and (min-width: $bp-sm-max) {
