+++ frontend/src/components/mobile/index.js
@@ -1,0 +1,1 @@
+export { default as MobileNav } from "./NavBarLinks.vue";
